backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
Pillow==10.4.0
supabase==2.4.2
python-dotenv==1.0.1
pytesseract==0.3.10
